+++ RELEASES.md
@@ -5,3 +5,4 @@
 - [#1](https://github.com/bvfnbk/planning-poker/issues/1) Created "Frontend" application module.
 - [#5](https://github.com/bvfnbk/planning-poker/issues/5) Serving "Frontend" application module using a Container.
 - [#3](https://github.com/bvfnbk/planning-poker/issues/3) Created "Gateway" application module.
+- #2 Created "Gateway" application module.